// packages/free-use-bible-api/models/ChapterVerse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Free Use Bible API
 * An API for accessing over 1000 different Bible translations.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: hello@helloao.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ChapterVerseContentInner } from './ChapterVerseContentInner';
import {
    ChapterVerseContentInnerFromJSON,
    ChapterVerseContentInnerFromJSONTyped,
    ChapterVerseContentInnerToJSON,
    ChapterVerseContentInnerToJSONTyped,
} from './ChapterVerseContentInner';

/**
 * Defines the schema for a verse in a chapter.
 * @export
 * @interface ChapterVerse
 */
export interface ChapterVerse {
    /**
     * 
     * @type {ChapterVerseTypeEnum}
     * @memberof ChapterVerse
     */
    type: ChapterVerseTypeEnum;
    /**
     * The number of the verse.
     * @type {number}
     * @memberof ChapterVerse
     */
    number: number;
    /**
     * The list of content for the verse. Each element in the list could be a string, formatted text, or a footnote reference.
     * @type {Array<ChapterVerseContentInner>}
     * @memberof ChapterVerse
     */
    content: Array<ChapterVerseContentInner>;
}


/**
 * @export
 */
export const ChapterVerseTypeEnum = {
    Verse: 'verse'
} as const;
export type ChapterVerseTypeEnum = typeof ChapterVerseTypeEnum[keyof typeof ChapterVerseTypeEnum];


/**
 * Check if a given object implements the ChapterVerse interface.
 */
export function instanceOfChapterVerse(value: object): value is ChapterVerse {
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('number' in value) || value['number'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    return true;
}

export function ChapterVerseFromJSON(json: any): ChapterVerse {
    return ChapterVerseFromJSONTyped(json, false);
}

export function ChapterVerseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChapterVerse {
    if (json == null) {
        return json;
    }
    return {
        
        'type': json['type'],
        'number': json['number'],
        'content': ((json['content'] as Array<any>).map(ChapterVerseContentInnerFromJSON)),
    };
}

export function ChapterVerseToJSON(json: any): ChapterVerse {
    return ChapterVerseToJSONTyped(json, false);
}

export function ChapterVerseToJSONTyped(value?: ChapterVerse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'type': value['type'],
        'number': value['number'],
        'content': ((value['content'] as Array<any>).map(ChapterVerseContentInnerToJSON)),
    };
}